+++ src/main/java/org/openlmis/requisition/web/RequisitionV2Controller.java
@@ -42,8 +42,11 @@ import org.openlmis.requisition.dto.OrderableDto;
 import org.openlmis.requisition.dto.RequisitionLineItemV2Dto;
 import org.openlmis.requisition.dto.RequisitionV2Dto;
 import org.openlmis.requisition.dto.VersionObjectReferenceDto;
+import org.openlmis.requisition.service.RequisitionService;
 import org.slf4j.profiler.Profiler;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.i18n.LocaleContextHolder;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.transaction.annotation.Transactional;
@@ -61,6 +64,9 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping(RESOURCE_URL)
 public class RequisitionV2Controller extends BaseRequisitionController {
+
+  @Autowired
+  private RequisitionService requisitionService;
 
   public static final String RESOURCE_URL = API_URL + "/v2/requisitions";
 
@@ -126,6 +132,9 @@ public class RequisitionV2Controller extends BaseRequisitionController {
     requisitionToUpdate.updateFrom(result.getRequisition(),
         result.getOrderables(), result.getApprovedProducts(),
         datePhysicalStockCountCompletedEnabledPredicate.exec(result.getProgram()));
+
+    requisitionService.processUnSkippedRequisitionLineItems(requisitionToUpdate,
+            LocaleContextHolder.getLocale());
 
     profiler.start("SAVE");
     requisitionRepository.save(requisitionToUpdate);